ESP32 Radar: Replace side arrows with cardinal indicator
The 1-3 arrows that hang off the side of the ESP32 Radar POIs are not intuitive for finding your flying buddies. This PR replaces the 1-3 arrows with a compass style cardinal indicator. This will point in the direction of the other craft. It is displayed on the right of the POI, to be distinct from the relative heading arrow.